Fix #195: Separate platform specific configuration source implementations into their own packages
