+++ src/Microsoft.Framework.ConfigurationModel.Ini/Properties/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Reflection;
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Microsoft.Framework.ConfigurationModel.Ini.Test")]
+[assembly: AssemblyMetadata("Serviceable", "True")]
