+++ src/Microsoft.Framework.ConfigurationModel.Ini/Resources.resx
@@ -117,22 +117,16 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="Error_CannotActivateAbstractOrInterface" xml:space="preserve">
-    <value>Cannot create instance of type '{0}' because it is either abstract or an interface.</value>
+  <data name="Error_FileNotFound" xml:space="preserve">
+    <value>The configuration file '{0}' was not found and is not optional.</value>
   </data>
-  <data name="Error_FailedBinding" xml:space="preserve">
-    <value>Failed to convert '{0}' to type '{1}'.</value>
+  <data name="Error_InvalidFilePath" xml:space="preserve">
+    <value>File path must be a non-empty string.</value>
   </data>
-  <data name="Error_FailedToActivate" xml:space="preserve">
-    <value>Failed to create instance of type '{0}'.</value>
+  <data name="Error_KeyIsDuplicated" xml:space="preserve">
+    <value>A duplicate key '{0}' was found.</value>
   </data>
-  <data name="Error_MissingBasePath" xml:space="preserve">
-    <value>Unable to resolve path '{0}'; construct this {1} with a non-null {2}.</value>
-  </data>
-  <data name="Error_MissingParameterlessConstructor" xml:space="preserve">
-    <value>Cannot create instance of type '{0}' because it is missing a public parameterless constructor.</value>
-  </data>
-  <data name="Error_NoSources" xml:space="preserve">
-    <value>A configuration source is not registered. Please register one before setting a value.</value>
+  <data name="Error_UnrecognizedLineFormat" xml:space="preserve">
+    <value>Unrecognized line format: '{0}'.</value>
   </data>
 </root>
