+++ test/Microsoft.Framework.ConfigurationModel.CommandLine.Test/project.json
@@ -1,10 +1,12 @@
-{
+﻿{
   "dependencies": {
-    "Microsoft.Framework.ConfigurationModel.Ini": "1.0.0-*",
-    "Microsoft.Framework.ConfigurationModel.Json": "1.0.0-*",
+    "Microsoft.Framework.ConfigurationModel.CommandLine": "1.0.0-*",
     "Microsoft.Framework.ConfigurationModel.Test.Common": "1.0.0-*",
-    "Microsoft.Framework.ConfigurationModel.Xml": "1.0.0-*",
     "xunit.runner.aspnet": "2.0.0-aspnet-*"
+  },
+  "frameworks": {
+    "dnx451": { },
+    "dnxcore50": { }
   },
   "commands": {
     "test": "xunit.runner.aspnet"
