+++ test/Microsoft.Framework.ConfigurationModel.FunctionalTests/ConfigurationTests.cs
@@ -4,7 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
+using Microsoft.Framework.ConfigurationModel.Ini;
 using Xunit;
 
 namespace Microsoft.Framework.ConfigurationModel.Test
